it's missing few information about routes for simplicity sake
the documentation are great but it's missing information about routes and how do we submit our forms, i have to admit that you've made it kinda scary as developers have to work in many files to get the form ready.
consider making it in simple steps and provide full documentation for a working form example like you've been doing with the controller part. 
one last thing i was wondering if a developer wants to add their own forms like HTML form without doing it this way? wouldn't be better to make that available? thank you for your time :)